+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 ### Added
 
 * Support for PipeWire, using its pulseaudio interface. Echo-cancellation is disabled when using PipeWire, since it doesn't support it yet.
+* New `--no-audio/-n` option to disable audio setup.
 
 ## [0.1.1] - 2021-04-23